Added constraint-based citations.
\cite{!author=sirer:title=hybrid-routing:year=2004}
will cite the unique paper whose author, title and year contain the
substrings shown. Any field can be specified as a constraint in the
citations. If the first field name is left out, it defaults to author.
If the second field is numeric and name is left out, defaults to year.
If the third one is left out, it defaults to year. 

So \cite{!sirer:sharp-routing:2004} is synonymous with the above,
so is \cite{!sirer:2004}.

Crosstex will complain if the same object is cited under multiple
constraint-based citations. It will also complain if the constraints
are non-unique given the database.

I lifted the idea in its entireity from Norman Ramseys nbibtex. Not
sure if it will pan out, but it seems like a clean interface.

git-svn-id: https://crosstex.svn.sourceforge.net/svnroot/crosstex/trunk@67 bdde6707-1129-0410-9c4d-a4f2bc757222